Use custom values to specify egress rules for services
Since we require the sso adress to be dynamic, but onyxia doesn't allow replacing just one value in the element.
The values are therefore provided as customValues to onyxia-api, which we then can use in onyxia-web to replace them in the chart.
Another approach would be to introduce a new variable in the chart for "static" egress rules and combine the two variables in the manifest where the serviceentries are created.

Internal ref: [DPSKY-498]